+++ index.d.ts
@@ -138,9 +138,14 @@ export declare interface RestrictionSettingsData {
 export declare class Account {
   constructor(config: Record<string, unknown>);
 
-  getToken(): Credentials;
-  setToken(token: string, refreshToken: string, expiresAt: number): void;
-  removeToken(): void;
+  isAuthenticated<R = boolean>(): R | boolean;
+  getToken<R = Credentials>(): R | Credentials;
+  setToken<R = void>(
+    token: string,
+    refreshToken: string,
+    expiresAt: number,
+  ): R | void;
+  removeToken<R = void>(): R | void;
 
   signIn(data: AuthenticateData): Promise<AxiosResponse<CreateAccount>>;
   signUp(data: SignUpData): Promise<AxiosResponse<CreateAccount>>;
@@ -248,6 +253,24 @@ export declare interface AccessType {
   period: string;
   updated_at: number;
   created_at: number;
+}
+export interface DonationOption {
+  id: number;
+  item_id: number;
+  amount: number;
+  currency: string;
+  description?: string;
+}
+
+export interface CustomDonationOption {
+  id: number;
+  item_id: number;
+  custom_price_enabled: boolean;
+}
+
+export interface DonationDetails {
+  donations: Array<DonationOption> | null;
+  donation_options: CustomDonationOption;
 }
 
 export declare interface TrialPeriod {
@@ -395,6 +418,7 @@ export declare class Asset {
   ): Promise<AxiosResponse<ExternalItemDetails>>;
   getPackage(id: number): Promise<AxiosResponse<GetMerchantPackage>>;
   getAssetAccessFees(id: number): Promise<AxiosResponse<GetAccessFee>>;
+  getDonationOptions(assetId: number): Promise<AxiosResponse<DonationDetails>>;
   getAssetsHistory(
     size?: number,
     page?: number,
@@ -475,6 +499,29 @@ export interface CreatePaymentData {
   brandingId?: number;
   receiverEmail?: string;
   isGift?: boolean;
+}
+
+export interface CreateDonationPaymentData {
+  number: number;
+  cardName: string;
+  expMonth: string;
+  expYear: number;
+  cvv: number;
+  paymentMethod: string;
+  referrer: string;
+  brandingId: number;
+  returnUrl: string;
+  amount: number;
+  currency: string;
+  assetId: number;
+  donationId?: number;
+}
+
+export interface ConfirmDonationPaymentData {
+  paymentIntentId: string;
+  brandingId: number;
+  paymentMethod: string;
+  donationId?: number;
 }
 
 export interface PayPalParamsData {
@@ -677,6 +724,7 @@ export declare class Payment {
   getPaymentMethods(): Promise<AxiosResponse<MerchantPaymentMethod[]>>;
   getPaymentTools(paymentMethodId: number): Promise<AxiosResponse<any>>;
   createPayment(data: CreatePaymentData): Promise<AxiosResponse<CreatePayment>>;
+  createDonationPayment(data: CreateDonationPaymentData): Promise<AxiosResponse<CreatePayment>>;
   getPayPalParams(
     data: PayPalParamsData
   ): Promise<AxiosResponse<GeneratePayPalParameters>>;
@@ -703,6 +751,9 @@ export declare class Payment {
   ) => Promise<AxiosResponse<CommonResponse>>;
   confirmPayment(
     paymentIntentId: string
+  ): Promise<AxiosResponse<CreatePayment>>;
+  confirmDonationPayment(
+    data: ConfirmDonationPaymentData
   ): Promise<AxiosResponse<CreatePayment>>;
   idealPayment: (
     data: IdealPaymentData
@@ -901,9 +952,9 @@ export declare class Notifications {
 type Env = 'development' | 'production';
 
 export interface LocalStorageMethods {
-  setItem: (key: string, value: string) => void;
-  getItem: (key: string) => string | null;
-  removeItem: (key: string) => void;
+  setItem: <R = void>(key: string, value: string) => R | void;
+  getItem: <R = string | null>(key: string) => R | (string | null);
+  removeItem: <R = void>(key: string) => R | void;
 }
 
 export type TokenStorageType = LocalStorageMethods & {