+++ src/endpoints/asset.ts
@@ -10,6 +10,7 @@ import BaseExtend from '../extends/base';
 import { API } from '../constants';
 import { CommonResponse } from '../models/CommonInterfaces';
 import tokenStorage from '../factories/tokenStorage';
+import { isPromise } from '../helpers';
 
 /**
  * Contains all Requests connected with assets/items
@@ -32,9 +33,11 @@ class Asset extends BaseExtend {
    * @returns  {AxiosResponse<GetItemAccessV1>}
    */
   async checkAccessForAsset(id: number) {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(API.checkAccessForAsset(id), {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -51,9 +54,11 @@ class Asset extends BaseExtend {
    * @returns  {AxiosResponse<boolean>}
    */
   async isFreeTrialUsed(id: number) {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(API.checkFreeTrial(id), {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -149,11 +154,13 @@ class Asset extends BaseExtend {
     endDate?: string,
     type?: string,
   ) {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(
       API.getAssetsHistory(size, page, startDate, endDate, type),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
         },
       },
     );
@@ -200,7 +207,7 @@ class Asset extends BaseExtend {
       browser_fingerprint: browserFingerprint,
     };
 
-    tokenStorage.setItem(
+    await tokenStorage.setItem(
       this.config.INPLAYER_ACCESS_CODE_NAME(codeAccessData.item_id),
       JSON.stringify(accessCode),
     );
@@ -252,11 +259,12 @@ class Asset extends BaseExtend {
       this.config.INPLAYER_ACCESS_CODE_NAME(assetId),
     );
 
-    if (!accessCode) {
-      return null;
+    if (isPromise(accessCode)) {
+      return (accessCode as Promise<string>).then((resolvedString) =>
+        (resolvedString ? (JSON.parse(resolvedString) as CodeAccessData) : null)) as Promise<CodeAccessData | null>;
     }
 
-    return JSON.parse(accessCode);
+    return accessCode ? (JSON.parse(accessCode as string) as CodeAccessData) : null;
   }
 
   /**
@@ -272,7 +280,7 @@ class Asset extends BaseExtend {
    * @returns  {Object}
    */
   async releaseAccessCode(assetId: number) {
-    const accessCode: CodeAccessData = this.getAccessCode(assetId);
+    const accessCode: CodeAccessData | null = await this.getAccessCode(assetId);
 
     if (!accessCode) {
       return null;
@@ -288,7 +296,9 @@ class Asset extends BaseExtend {
       { data: formData },
     );
 
-    tokenStorage.removeItem(this.config.INPLAYER_ACCESS_CODE_NAME(assetId));
+    await tokenStorage.removeItem(
+      this.config.INPLAYER_ACCESS_CODE_NAME(assetId),
+    );
 
     return response;
   }
@@ -305,7 +315,7 @@ class Asset extends BaseExtend {
    * @returns  null
    */
   async terminateSession(assetId: number) {
-    const accessCode: CodeAccessData = this.getAccessCode(assetId);
+    const accessCode: CodeAccessData | null = await this.getAccessCode(assetId);
 
     if (!accessCode) {
       return null;
@@ -315,7 +325,9 @@ class Asset extends BaseExtend {
       API.terminateSession(accessCode.code, accessCode.browser_fingerprint),
     );
 
-    tokenStorage.removeItem(this.config.INPLAYER_ACCESS_CODE_NAME(assetId));
+    await tokenStorage.removeItem(
+      this.config.INPLAYER_ACCESS_CODE_NAME(assetId),
+    );
 
     return response;
   }
@@ -335,9 +347,32 @@ class Asset extends BaseExtend {
    * }
    */
   async getCloudfrontURL(assetId: number, videoUrl: string) {
+    const tokenObject = await this.request.getToken();
+
     return this.request.get(API.getCloudfrontURL(assetId, videoUrl), {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
+      },
+    });
+  }
+
+  /**
+   * Gets the donation options for the asset.
+   * @method getDonationOptions
+   * @async
+   * @param {number} assetId - The id of the asset
+   * @example
+   *     InPlayer.Donation
+   *     .getDonationOptions(42597)
+   *     .then(data => console.log(data));
+   * @returns {AxiosResponse<DonationDetails>}
+   */
+  async getDonationOptions(assetId: number) {
+    const tokenObject = await this.request.getToken();
+
+    return this.request.get(API.getDonations(assetId), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }