+++ src/endpoints/payment.ts
@@ -12,6 +12,10 @@ import {
   GoogleOrApplePaymentRequestBody,
   ValidateReceiptData,
   ReceiptValidationPlatform,
+  CreateDonationPaymentData,
+  CreateDonationPaymentRequestBody,
+  ConfirmDonationPaymentData,
+  ConfirmDonationPaymentRequestBody,
   StripePaymentPaymentMethods,
 } from '../models/IPayment&Subscription';
 import { CustomErrorResponse } from '../models/CommonInterfaces';
@@ -41,9 +45,11 @@ class Payment extends BaseExtend {
    * @returns  {AxiosResponse<Array<MerchantPaymentMethod>>}
    */
   async getPaymentMethods() {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(API.getPaymentMethods, {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -60,14 +66,13 @@ class Payment extends BaseExtend {
    * @returns  {AxiosResponse<any>}
    */
   async getPaymentTools(paymentMethodId: number) {
-    return this.request.authenticatedGet(
-      API.getPaymentTools(paymentMethodId),
-      {
-        headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
-        },
+    const tokenObject = await this.request.getToken();
+
+    return this.request.authenticatedGet(API.getPaymentTools(paymentMethodId), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
       },
-    );
+    });
   }
 
   /**
@@ -132,12 +137,80 @@ class Payment extends BaseExtend {
       body.receiver_email = data.receiverEmail;
     }
 
+    const tokenObject = await this.request.getToken();
+
+    return this.request.authenticatedPost(API.payForAsset, qs.stringify(body), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
+    });
+  }
+
+  /**
+   * Makes a Donation Payment for a given Authorization token + asset/payment details.
+   * @method createDonationPayment
+   * @async
+   * @param {Object} data - Payment data - {
+   *  number: number || string,
+   *  cardName: string,
+   *  expMonth: number,
+   *  expYear: number,
+   *  cvv: number,
+   *  assetId: number,
+   *  paymentMethod: string,
+   *  referrer: string,
+   *  brandingId?: number,
+   *  amount: number,
+   *  currency: string,
+   *  donationId: number,
+   * }
+   * @example
+   *     InPlayer.Payment
+   *     .createPayment({
+   *       number: 4111111111111111,
+   *       cardName: 'Test',
+   *       expMonth: 10,
+   *       expYear: 2030,
+   *       cvv: 656,
+   *       assetId: 2341,
+   *       paymentMethod: 'Card',
+   *       referrer: 'http://google.com',
+   *       voucherCode: 'fgh1982gff-0f2grfds'
+   *       brandingId: 1234,
+   *       amount: 1234,
+   *       currency: EUR,
+   *       donationId: 4567,
+   *       returnUrl: 'https://event.inplayer.com/staging',
+   *      })
+   *     .then(data => console.log(data));
+   * @returns  {AxiosResponse<CreateDonationPayment>}
+   */
+  async createDonationPayment(data: CreateDonationPaymentData) {
+    const body: CreateDonationPaymentRequestBody = {
+      number: data.number,
+      card_name: data.cardName,
+      exp_month: data.expMonth,
+      exp_year: data.expYear,
+      cvv: data.cvv,
+      payment_method: data.paymentMethod,
+      referrer: data.referrer,
+      branding_id: data.brandingId,
+      currency_iso: data.currency,
+      amount: data.amount,
+      item_id: data.assetId,
+      donation_id: data.donationId,
+      return_url: buildURLwithQueryParams(data.returnUrl, { ippwat: 'ppv' }),
+    };
+
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
-      API.payForAsset,
+      API.payForAssetDonation,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -175,12 +248,64 @@ class Payment extends BaseExtend {
       pi_id: paymentIntentId,
     };
 
+    const tokenObject = await this.request.getToken();
+
+    return this.request.authenticatedPost(API.payForAsset, qs.stringify(body), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
+    });
+  }
+
+  /**
+   * As part of new bank regulations, we need to provide options for
+   * additional authentication during the payment flow for customers
+   * @method confirmDonationPayment
+   * @async
+   * @param {string}
+   * @example
+   *     InPlayer.Payment
+   *     .confirmDonationPayment('332242', 1, 'Card')
+   *     .then(data => console.log(data));
+   * @returns  {AxiosResponse<CreatePayment>} Contains the data - {
+   *       message: "Submitted for payment",
+   *  }
+   */
+  async confirmDonationPayment(data: ConfirmDonationPaymentData) {
+    const {
+      paymentIntentId,
+      brandingId,
+      paymentMethod,
+      donationId,
+    } = data;
+    if (!paymentIntentId) {
+      const response: CustomErrorResponse = {
+        status: 400,
+        data: {
+          code: 400,
+          message: 'Payment Intend Id is a required parameter!',
+        },
+      };
+
+      throw { response };
+    }
+
+    const body: ConfirmDonationPaymentRequestBody = {
+      pi_id: paymentIntentId,
+      branding_id: brandingId,
+      payment_method: paymentMethod,
+      donation_id: donationId,
+    };
+
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
-      API.payForAsset,
+      API.confirmForAssetDonation,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -220,9 +345,11 @@ class Payment extends BaseExtend {
       formData.append('voucher_code', data.voucherCode);
     }
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(API.getPayPalParams, formData, {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -241,11 +368,13 @@ class Payment extends BaseExtend {
    * @returns  {AxiosResponse<PurchaseHistoryCollection[]>}
    */
   async getPurchaseHistory(status = 'active', page: number, limit: number) {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(
       API.getPurchaseHistory(status, page, limit),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
         },
       },
     );
@@ -262,9 +391,11 @@ class Payment extends BaseExtend {
    * @returns  {AxiosResponse<GetDefaultCard>}
    */
   async getDefaultCreditCard() {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(API.getDefaultCreditCard, {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -304,12 +435,14 @@ class Payment extends BaseExtend {
       currency_iso: data.currency,
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPut(
       API.setDefaultCreditCard,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -339,9 +472,11 @@ class Payment extends BaseExtend {
    * }
    */
   async getDirectDebitMandate() {
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedGet(API.getDirectDebitMandate, {
       headers: {
-        Authorization: `Bearer ${this.request.getToken().token}`,
+        Authorization: `Bearer ${tokenObject.token}`,
       },
     });
   }
@@ -384,12 +519,14 @@ class Payment extends BaseExtend {
       iban: data.iban,
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.createDirectDebitMandate,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -426,12 +563,14 @@ class Payment extends BaseExtend {
       referrer: data.referrer,
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.payForAssetV2,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -468,16 +607,14 @@ class Payment extends BaseExtend {
       referrer: data.referrer,
     };
 
-    return this.request.authenticatedPost(
-      API.subscribeV2,
-      qs.stringify(body),
-      {
-        headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
+    const tokenObject = await this.request.getToken();
+
+    return this.request.authenticatedPost(API.subscribeV2, qs.stringify(body), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
+        'Content-Type': 'application/x-www-form-urlencoded',
       },
-    );
+    });
   }
 
   /**
@@ -524,12 +661,14 @@ class Payment extends BaseExtend {
       body.voucher_code = data.voucherCode;
     }
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.payForAssetV2,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -575,12 +714,14 @@ class Payment extends BaseExtend {
       referrer,
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.payForAssetV2,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },
@@ -626,12 +767,14 @@ class Payment extends BaseExtend {
       referrer,
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.payForAssetV2,
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       }
@@ -682,16 +825,14 @@ class Payment extends BaseExtend {
       body.voucher_code = data.voucherCode;
     }
 
-    return this.request.authenticatedPost(
-      API.subscribeV2,
-      qs.stringify(body),
-      {
-        headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
+    const tokenObject = await this.request.getToken();
+
+    return this.request.authenticatedPost(API.subscribeV2, qs.stringify(body), {
+      headers: {
+        Authorization: `Bearer ${tokenObject.token}`,
+        'Content-Type': 'application/x-www-form-urlencoded',
       },
-    );
+    });
   }
 
   /**
@@ -734,12 +875,14 @@ class Payment extends BaseExtend {
         : {}),
     };
 
+    const tokenObject = await this.request.getToken();
+
     return this.request.authenticatedPost(
       API.validateReceipt(String(platform)),
       qs.stringify(body),
       {
         headers: {
-          Authorization: `Bearer ${this.request.getToken().token}`,
+          Authorization: `Bearer ${tokenObject.token}`,
           'Content-Type': 'application/x-www-form-urlencoded',
         },
       },